backend/requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
mypy==1.16.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.2
pyright==1.1.403
pytest==8.4.1
pytz==2025.2
PyYAML==6.0.2
ruff==0.12.2
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
